refactor(HDF5): moved template functions to cpp files
